+++ packages/luciq_flutter/lib/src/utils/screen_loading/screen_loading_manager.dart
@@ -4,9 +4,9 @@ import 'package:luciq_flutter/src/utils/lcq_build_info.dart';
 import 'package:luciq_flutter/src/utils/lcq_date_time.dart';
 import 'package:luciq_flutter/src/utils/luciq_logger.dart';
 import 'package:luciq_flutter/src/utils/luciq_montonic_clock.dart';
-import 'package:luciq_flutter/src/utils/screen_loading/flags_config.dart';
 import 'package:luciq_flutter/src/utils/screen_loading/screen_loading_trace.dart';
 import 'package:luciq_flutter/src/utils/screen_loading/ui_trace.dart';
+import 'package:luciq_flutter/src/utils/ui_trace/flags_config.dart';
 import 'package:meta/meta.dart';
 
 /// Manages screen loading traces and UI traces for performance monitoring.
@@ -139,7 +139,7 @@ class ScreenLoadingManager {
   /// [matchingScreenName] as the screen name used for matching the UI trace
   /// with a Screen Loading trace.
   @internal
-  Future<void> startUiTrace(
+  Future<int?> startUiTrace(
     String screenName, [
     String? matchingScreenName,
   ]) async {
@@ -150,21 +150,19 @@ class ScreenLoadingManager {
 
       final isSDKBuilt =
           await _checkLuciqSDKBuilt("APM.LuciqCaptureScreenLoading");
-      if (!isSDKBuilt) return;
 
-      // TODO: On Android, FlagsConfig.apm.isEnabled isn't implemented correctly
-      // so we skip the isApmEnabled check on Android and only check on iOS.
-      // This is a temporary fix until we implement the isEnabled check correctly.
-      // We need to fix this in the future.
-      final isApmEnabled = await FlagsConfig.apm.isEnabled();
-      if (!isApmEnabled && LCQBuildInfo.I.isIOS) {
+      if (!isSDKBuilt) return null;
+
+      final isAutoUiTraceEnabled = await FlagsConfig.uiTrace.isEnabled();
+
+      if (!isAutoUiTraceEnabled) {
         LuciqLogger.I.e(
-          'APM is disabled, skipping starting the UI trace for screen: $screenName.\n'
+          'Auto UI trace is disabled, skipping starting the UI trace for screen: $screenName.\n'
           'Please refer to the documentation for how to enable APM on your app: '
           'https://docs.luciq.ai/docs/react-native-apm-disabling-enabling',
           tag: APM.tag,
         );
-        return;
+        return null;
       }
 
       final sanitizedScreenName = sanitizeScreenName(screenName);
@@ -181,8 +179,10 @@ class ScreenLoadingManager {
         matchingScreenName: sanitizedMatchingScreenName,
         traceId: uiTraceId,
       );
+      return uiTraceId;
     } catch (error, stackTrace) {
       _logExceptionErrorAndStackTrace(error, stackTrace);
+      return null;
     }
   }