+++ packages/luciq_flutter/test/utils/screen_loading/screen_loading_manager_test.dart
@@ -7,12 +7,12 @@ import 'package:luciq_flutter/src/utils/lcq_build_info.dart';
 import 'package:luciq_flutter/src/utils/lcq_date_time.dart';
 import 'package:luciq_flutter/src/utils/luciq_logger.dart';
 import 'package:luciq_flutter/src/utils/luciq_montonic_clock.dart';
-import 'package:luciq_flutter/src/utils/screen_loading/flags_config.dart';
-import 'package:luciq_flutter/src/utils/screen_loading/screen_loading_manager.dart';
 import 'package:luciq_flutter/src/utils/screen_loading/screen_loading_trace.dart';
 import 'package:luciq_flutter/src/utils/screen_loading/ui_trace.dart';
+import 'package:luciq_flutter/src/utils/ui_trace/flags_config.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
+
 import 'screen_loading_manager_test.mocks.dart';
 
 class ScreenLoadingManagerNoResets extends ScreenLoadingManager {
@@ -186,11 +186,9 @@ void main() {
       verifyNever(mApmHost.startCpUiTrace(any, any, any));
     });
 
-    test('[startUiTrace] with APM disabled on iOS Platform should Log error',
-        () async {
+    test('[startUiTrace] with APM disabled should Log error', () async {
       mScreenLoadingManager.currentUiTrace = uiTrace;
-      when(FlagsConfig.apm.isEnabled()).thenAnswer((_) async => false);
-      when(LCQBuildInfo.I.isIOS).thenReturn(true);
+      when(FlagsConfig.uiTrace.isEnabled()).thenAnswer((_) async => false);
 
       await ScreenLoadingManager.I.startUiTrace(screenName);
 
@@ -199,7 +197,7 @@ void main() {
 
       verify(
         mLuciqLogger.e(
-          'APM is disabled, skipping starting the UI trace for screen: $screenName.\n'
+          'Auto UI trace is disabled, skipping starting the UI trace for screen: $screenName.\n'
           'Please refer to the documentation for how to enable APM on your app: https://docs.luciq.ai/docs/react-native-apm-disabling-enabling',
           tag: APM.tag,
         ),
@@ -210,8 +208,7 @@ void main() {
     test(
         '[startUiTrace] with APM enabled on android Platform should call `APM.startCpUiTrace and set UiTrace',
         () async {
-      when(FlagsConfig.apm.isEnabled()).thenAnswer((_) async => true);
-      when(LCQBuildInfo.I.isIOS).thenReturn(false);
+      when(FlagsConfig.uiTrace.isEnabled()).thenAnswer((_) async => true);
 
       await ScreenLoadingManager.I.startUiTrace(screenName);
 
@@ -235,8 +232,7 @@ void main() {
     test(
         '[startUiTrace] with APM enabled should create a UI trace with the matching screen name',
         () async {
-      when(FlagsConfig.apm.isEnabled()).thenAnswer((_) async => true);
-      when(LCQBuildInfo.I.isIOS).thenReturn(false);
+      when(FlagsConfig.uiTrace.isEnabled()).thenAnswer((_) async => true);
       when(
         RouteMatcher.I.match(
           routePath: anyNamed('routePath'),