+++ packages/luciq_flutter/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Changelog
 
+## [19.2.2] (https://github.com/luciqai/luciq-flutter-sdk/compare/v19.2.2...19.2.1) (February 24, 2026)
+
+### Changed
+
+- Bump Luciq iOS SDK to v19.4.1 ([#41](https://github.com/luciqai/luciq-flutter-sdk/pull/41)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/19.4.1).
+
+- Bump Luciq Android SDK to v19.2.2 ([#41](https://github.com/luciqai/luciq-flutter-sdk/pull/41)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v19.4.1).
+
+
+## [19.2.1] (https://github.com/luciqai/luciq-flutter-sdk/compare/v19.2.1...19.1.0) (January 28, 2026)
+
+### Added
+
+- Add new UserConsentActionType - noAutomaticBugGrouping.
+
+### Changed
+
+- Bump Luciq iOS SDK to v19.4.0 ([#36](https://github.com/luciqai/luciq-flutter-sdk/pull/36)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/19.4.0).
+
+- Bump Luciq Android SDK to v19.2.1 ([#36](https://github.com/luciqai/luciq-flutter-sdk/pull/36)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v19.2.1).
+
+
+## [19.1.0] (https://github.com/luciqai/luciq-flutter-sdk/compare/v19.1.0...19.0.0) (January 5, 2026)
+
+### Changed
+
+- Bump Luciq iOS SDK to v19.3.0 ([#33](https://github.com/luciqai/luciq-flutter-sdk/pull/33)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/19.3.0).
+
+- Bump Luciq Android SDK to v19.1.0 ([#33](https://github.com/luciqai/luciq-flutter-sdk/pull/33)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v19.1.0).
+
+## [19.0.0] (https://github.com/luciqai/luciq-flutter-sdk/compare/v19.0.0...18.2.1) (December 16, 2025)
+
+### Added
+
+- Add support for Screen Rendering ([#22](https://github.com/luciqai/luciq-flutter-sdk/pull/22))
+
+### Changed
+
+- Bump Luciq iOS SDK to v19.2.0 ([#22](https://github.com/luciqai/luciq-flutter-sdk/pull/10)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/19.2.0).
+
+- Bump Luciq Android SDK to v19.0.0 ([#22](https://github.com/luciqai/luciq-flutter-sdk/pull/10)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v19.0.0).
+
+
 ## [18.2.1] (https://github.com/luciqai/luciq-flutter-sdk/compare/v18.2.1...18.2.0) (November 26, 2025)
 
 ### Fixed
@@ -18,7 +61,6 @@
 
 - Bump Luciq Android SDK to v18.2.0 ([#10](https://github.com/luciqai/luciq-flutter-sdk/pull/10)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v18.2.0).
 
-
 ## [18.0.1] (https://github.com/luciqai/luciq-flutter-sdk/compare/v18.0.1...18.0.0) (October 27, 2025)
 
 ### Added
@@ -34,7 +76,6 @@
 - Bump Luciq iOS SDK to v18.0.1 ([#10](https://github.com/luciqai/luciq-flutter-sdk/pull/10)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/18.0.1).
 
 - Bump Luciq Android SDK to v18.0.1 ([#10](https://github.com/luciqai/luciq-flutter-sdk/pull/10)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v18.0.1).
-
 
 ## [18.0.0](https://github.com/luciqai/luciq-flutter-sdk/compare/v18.0.0...dev) (September 24, 2025)